+++ app/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.k.hilaris.alpha">
+    package="com.l.hilaris.alpha">
 
     <uses-permission android:name="android.permission.INTERNET"/>
 
@@ -27,19 +27,19 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.k.hilaris.alpha.views.front.FrontActivity"
+            android:name="com.l.hilaris.alpha.views.front.FrontActivity"
             android:label="Main Menu">
         </activity>
         <activity
-            android:name="com.k.hilaris.alpha.views.singleplayer.SinglePlayerMenuActivity"
+            android:name="com.l.hilaris.alpha.views.singleplayer.SinglePlayerMenuActivity"
             android:label="Select Difficulty">
         </activity>
         <activity
-            android:name="com.k.hilaris.alpha.views.multiplayer.MultiPlayerMenuActivity"
+            android:name="com.l.hilaris.alpha.views.multiplayer.MultiPlayerMenuActivity"
             android:label="Online Session">
         </activity>
         <activity
-            android:name="com.k.hilaris.alpha.views.sudoku.singleplayer.SudokuActivity"
+            android:name="com.l.hilaris.alpha.views.sudoku.singleplayer.SudokuActivity"
             android:label="Sudoku">
         </activity>
     </application>
