+++ app/src/main/java/com/l/hilaris/alpha/models/SudokuVariation.java
@@ -1,17 +1,20 @@
 package com.l.hilaris.alpha.models;
 
+import java.io.Serializable;
 import java.util.List;
 
-public class SudokuVariation extends Sudoku {
+public class SudokuVariation extends Sudoku implements Serializable {
     public List<String> solution;
     public String guid;
+    public int score;
 
     public SudokuVariation() {
+        score = 0;
     }
 
     public SudokuVariation(Sudoku sudoku) {
-
         super(sudoku.getCells(), sudoku.getId(), sudoku.getDifficulty());
+        score = 0;
     }
 
     public List<String> getSolution() {
@@ -29,4 +32,13 @@ public class SudokuVariation extends Sudoku {
     public void setGuid(String guid) {
         this.guid = guid;
     }
+
+    public int getScore() {
+        return score;
+    }
+
+    public void setScore(int score) {
+        this.score = score;
+    }
+
 }
